socket-io-server/requirements.txt
python-socketio==4.6.1
eventlet==0.29.1